+++ R/object.r
@@ -8,8 +8,8 @@
 #'
 #' @name sc_object
 #' @export
-#' @seealso `sc_coord` for the coordinates part of the model, `sc_branch` for 
-#' the central part of the model, and `BRANCH` for the full model. 
+#' @seealso `sc_coord` for the coordinates part of the model, `sc_path` for 
+#' the central part of the model, and `PATH` for the full model. 
 sc_object <- function(x, ...) UseMethod("sc_object")
 
 #' @name sc_object