+++ NEWS.md
@@ -1,3 +1,11 @@
+# dev
+
+* revert to sf 0.2-7 to enable vignette migration
+* PATH model added (with ongoing migration from "branch" term to "path" term)
+* BRANCH model effectively removed, replaced by the PATH model
+* PRIMITIVES model added, including bare-nodes arc_node function to return the 3-way nodes
+
+
 # sc 0.0.1
 
 * first functions, the BRANCH model for sf